GRN: fix frontend lint and JS perf budget CI failures
Both failures pre-dated this branch (the isometric masterplan merged
with 11 lint errors and a bundle 12 KB over the 1 MiB total budget) but
they block this PR, so resolve them at the root:

Lint (react-hooks/refs + react-refresh):
- useMapViewport synced its latest-transform mirror ref during render;
  move the sync into an effect. That single render-time ref write was
  also what tainted every viewport property access in GardenMasterplan
  - destructuring the hook result clears the remaining reports
- Move bedFootprint, annotationFootprint, FLAT_KINDS, and sceneMetrics
  into a new footprints.ts so IsoBed/IsoAnnotation/IsoScene export only
  components, as react-refresh requires

Performance budget:
- Switch the Konva editor to react-konva's minimal-bundle pattern:
  import ReactKonvaCore plus only the shapes the designer renders
  instead of the full konva build (every shape + filters). Cuts the
  GardenDesignerPage chunk by 31.6 KB and brings the total to
  1,036,002 bytes, back under the 1,048,576-byte budget with headroom

Lint, typecheck, the 330-test suite, and the budget check all pass
locally.
