+++ apps/grn/src/components/GardenDesigner/AnnotationShape.tsx
@@ -1,7 +1,19 @@
 import type Konva from 'konva';
 import type { KonvaEventObject } from 'konva/lib/Node';
+import 'konva/lib/shapes/Circle';
+import 'konva/lib/shapes/Line';
+import 'konva/lib/shapes/Rect';
+import 'konva/lib/shapes/Text';
+import 'konva/lib/shapes/Transformer';
 import { useEffect, useRef } from 'react';
-import { Circle, Group, Line, Rect, Text, Transformer } from 'react-konva';
+import {
+  Circle,
+  Group,
+  Line,
+  Rect,
+  Text,
+  Transformer,
+} from 'react-konva/lib/ReactKonvaCore';
 import type { GardenAnnotation } from '../../types/listing';
 
 interface AnnotationShapeProps {
